added choose car() and askForTimeInSeconds() as well as the main game loop.

+++ Aufgabe6-Auto/BMW.h
@@ -16,8 +16,8 @@ namespace car {
     class BMW : public car::BaseCar {
     public:
         BMW();
-        void accelerate(int _time) override;
-        void deaccelerate(int _time) override;
+        void accelerate(float _time) override;
+        void deaccelerate(float _time) override;
         void brake() override;
     };
 }
