+++ Aufgabe6-Auto/Wheel.h
@@ -9,14 +9,14 @@ namespace car::component {
 
     class Wheel {
     public:
-        Wheel(std::string _name, int _powerTransmission) : name(_name), powerTransmission(_powerTransmission){};
+        Wheel(std::string _name, float _powerTransmission) : name(_name), powerTransmission(_powerTransmission){};
 
         std::string getName();
         int getPowerTransmission();
 
     protected:
         std::string name;
-        int powerTransmission;
+        float powerTransmission;
     };
 }
